% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_corner_label.R
\name{add_corner_label}
\alias{add_corner_label}
\title{Simple function for placing text labels in the corner of a figure.}
\usage{
add_corner_label(txt, where, inset = 0.08, inset.x = inset,
  inset.y = inset, log.y = FALSE, log.x = FALSE, ...)
}
\description{
Designed for labelling figures in the corner with 'A', 'B', and so on. Finds the coordinates for you.
}
\examples{

o <- par(mfrow=c(1,2))
with(iris, {
  plot(Sepal.Width, Sepal.Length, panel.last=add_corner_label("A","topleft"))
  plot(Petal.Width, Petal.Length, panel.last=add_corner_label("B","topleft"))
  })
par(o)
}

