% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_title.R
\name{axis_title}
\alias{axis_title}
\title{Adds an axis title in the margin}
\usage{
axis_title(text, side, ...)
}
\description{
A simple wrapper for \code{\link{mtext}}, making sure that the character size is adjusted to be identical to the character size used on the ordinary axes, and respecting the setting for \code{mgp} in \code{\link{par}}. This is especially handy when plotting multiple panels (via the use of \code{par(mfrow=...)}), because axis labels are automatically adjusted in that case (see entry for \code{mfrow} in \code{\link{par}}).
}
\examples{
# Plot a second axis
set.seed(1)
x <- rnorm(100)
y <- x + rnorm(100)
z <- rnorm(100) - x

o <- par(mar=c(5,5,2,5)) 
plot(x, y, ylab="X variable")
points(x, z, pch=19)
axis_title("Z variable", side = 4)
axis(4)
par(o)
}

