% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_bygroup.R
\name{hist_bygroup}
\alias{hist_bygroup}
\title{Plot and overlay histograms by group}
\usage{
hist_bygroup(yvar, group, dataset, breaks = 25, what = c("counts",
  "density"), log = FALSE, col = palette(), xlab = NULL,
  ylab = "Nr. individuals", ylim = NULL, xlim = NULL,
  legend.text = NULL, legend.cex = 1, xaxis = NULL, overlay = TRUE,
  group.expression = NULL, ...)
}
\description{
A simple function to plot histograms by group, and overlay them. This is notoriously difficult to achieve in base R.
}
\examples{
library(wesanderson)
hist_bygroup(Petal.Length, Species, iris, col=wes_palette("Chevalier"), breaks=50)

hist_bygroup(Petal.Length, Species, iris, col=wes_palette("Chevalier"), 
group.expression=abline(v=mean(Y)))
)
}

