% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideclipper-package.R
\docType{package}
\name{insideclipper-package}
\alias{insideclipper}
\alias{insideclipper-package}
\title{insideclipper: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diminutive/insideclipper}
  \item Report bugs at \url{https://github.com/diminutive/insideclipper/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
