% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10axes.R
\name{log10axes}
\alias{log10axes}
\title{Simple logarithmic axes function}
\usage{
log10axes(side = 1:2, logged = NULL, labels = TRUE)
}
\arguments{
\item{side}{Which side to add axes to (numbering as for \code{\link{axis}})}

\item{logged}{Which axes are already log-transformed (numbering as for \code{side})}

\item{labels}{If TRUE, add axis numeric labels}
}
\description{
Based on magaxis, with some modifications. Unlike \code{\link{magaxis}}, the labels are always of the form 10^i, where magaxis sometimes uses 10^i, other times it uses 10,100, etc., depending on the range.
}

