% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBy.R
\name{plotBy}
\alias{plotBy}
\title{Extension for formula-based plotting}
\usage{
plotBy(formula, data, how = c("colour", "panel"), fit = c("none", "gam",
  "lm"), ...)
}
\description{
Extension for formula-based plotting
}
\examples{
par(mfrow=c(1,2))
plotBy(log(Sepal.Length) + Petal.Length ~ Sepal.Width | Species, data=iris, fit="gam")

par(mfrow=c(1,2))
plotBy(log(Sepal.Length) + Petal.Length ~ Sepal.Width | Species, data=iris, fit="lm")

par(mfrow=c(1,2))
plotBy(log(Sepal.Length) + Petal.Length ~ Sepal.Width | Species, data=iris, how="colour")

par(mfrow=c(2,3))
plotBy(log(Sepal.Length) + Petal.Length ~ Sepal.Width | Species, data=iris, how="panel")


par(mfrow=c(1,1))
chickw <- subset(ChickWeight, !Chick \%in\% c("18","16"))
plotBy(weight ~ Time|Chick, fit="gam", data=chickw, kgam=5, linecols="black", band=F, lwd=1)

plotBy(weight ~ Time|Chick, type='l', data=chickw)
}

