% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plover-package.R
\docType{package}
\name{plover-package}
\alias{plover}
\alias{plover-package}
\title{plover: What the Package Does (One Line, Title Case)}
\description{
Point in polygon.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/plover}
  \item Report bugs at \url{https://github.com/hypertidy/plover/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
